requirements.txt
flask
boto3
requests
python-dotenv
google-genai
Pillow
gunicorn
